requirements.txt
Django==4.0.5
zappa==0.55.0
django-model-utils==4.2.0
pydash==5.1.0
django-cors-headers==3.11.0
mysqlclient==2.1.0
PyMySQL==1.0.2
pydash==5.1.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-spectacular==0.22.1
django-extensions==3.1.5
django-environ==0.8.1
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2